+++ Editor/API/ProjectAuditor.cs
@@ -157,6 +157,7 @@ namespace Unity.ProjectAuditor.Editor
 
             var logTimingsInfo = UserPreferences.LogTimingsInfo;
             var stopwatch = Stopwatch.StartNew();
+            var isCancelled = false;
             foreach (var module in supportedModules)
             {
                 var moduleStartTime = DateTime.Now;
@@ -168,25 +169,32 @@ namespace Unity.ProjectAuditor.Editor
                         report.AddIssues(resultsList);
                         analysisParams.OnIncomingIssues?.Invoke(resultsList);
                     },
-                    OnModuleCompleted = () =>
+                    OnModuleCompleted = (analysisResult) =>
                     {
                         var moduleEndTime = DateTime.Now;
+                        if (analysisResult == AnalysisResult.Cancelled)
+                            isCancelled = true;
+                        else if (analysisResult == AnalysisResult.Failure)
+                            Debug.Log($"[{ProjectAuditor.DisplayName}] Module {module.Name} failed.");
                         if (logTimingsInfo)
                             Debug.Log($"[{ProjectAuditor.DisplayName}] Module {module.Name} analysis took: " +
                                 (moduleEndTime - moduleStartTime).TotalMilliseconds / 1000.0 + " seconds.");
 
-                        report.RecordModuleInfo(module, moduleStartTime, moduleEndTime);
+                        report.RecordModuleInfo(module, moduleStartTime, moduleEndTime, analysisResult);
 
-                        analysisParams.OnModuleCompleted?.Invoke();
+                        analysisParams.OnModuleCompleted?.Invoke(analysisResult);
 
                         var finished = --numModules == 0;
                         if (finished)
                         {
                             stopwatch.Stop();
+                            if (isCancelled)
+                                Debug.Log($"[{ProjectAuditor.DisplayName}] Analysis was cancelled by the user.");
                             if (logTimingsInfo)
                                 Debug.Log($"[{ProjectAuditor.DisplayName}] Analysis took: " + stopwatch.ElapsedMilliseconds / 1000.0f +
                                     " seconds.");
 
+                            // finally, call the user's OnCompleted callback
                             analysisParams.OnCompleted?.Invoke(report);
                         }
                     }
@@ -194,12 +202,16 @@ namespace Unity.ProjectAuditor.Editor
 
                 try
                 {
-                    module.Audit(moduleParams, progress);
+                    var analysisResult = module.Audit(moduleParams, progress);
+                    if (analysisResult == AnalysisResult.Cancelled)
+                        progress.Clear();
+                    if (analysisResult != AnalysisResult.InProgress)
+                        moduleParams.OnModuleCompleted(analysisResult);
                 }
                 catch (Exception e)
                 {
                     Debug.LogError($"[{ProjectAuditor.DisplayName}] Module {module.Name} failed: " + e.Message + " " + e.StackTrace);
-                    moduleParams.OnModuleCompleted();
+                    moduleParams.OnModuleCompleted(AnalysisResult.Failure);
                 }
             }