+++ Editor/Modules/MeshModule.cs
@@ -49,7 +49,7 @@ namespace Unity.ProjectAuditor.Editor.Modules
             diagnosticParams.RegisterParameter(k_MeshTriangleCountLimit, 5000);
         }
 
-        public override void Audit(AnalysisParams analysisParams, IProgress progress = null)
+        public override AnalysisResult Audit(AnalysisParams analysisParams, IProgress progress = null)
         {
             var analyzers = GetPlatformAnalyzers(analysisParams.Platform);
 
@@ -71,6 +71,9 @@ namespace Unity.ProjectAuditor.Editor.Modules
 
             foreach (var assetPath in assetPaths)
             {
+                if (progress?.IsCancelled ?? false)
+                    return AnalysisResult.Cancelled;
+
                 context.Importer = AssetImporter.GetAtPath(assetPath);
 
                 foreach (var analyzer in analyzers)
@@ -83,7 +86,7 @@ namespace Unity.ProjectAuditor.Editor.Modules
 
             progress?.Clear();
 
-            analysisParams.OnModuleCompleted?.Invoke();
+            return AnalysisResult.Success;
         }
     }
 }